completed requirements for milestone-4
