+++ src/index.js
@@ -12,7 +12,7 @@ const initialState = {
     { id: getId(), title: 'The Last Flight', category: 'Action' },
     { id: getId(), title: 'Chronicles of Narnia', category: 'History' },
   ],
-  filter: "All"
+  filter: 'All',
 };
 
 const store = createStore(reducer, initialState);
@@ -23,4 +23,3 @@ ReactDOM.render(
   </Provider>,
   document.getElementById('root'),
 );
-
