+++ src/reducers/index.js
@@ -1,6 +1,8 @@
 import { combineReducers } from 'redux';
 import books from './books';
+import filter from './filter';
 
 export default combineReducers({
   books,
+  filter,
 });
